# api/openapi.yaml
openapi: 3.0.3
info:
  title: Golf Course Reservation Management System API
  description: 골프장 예약 관리 시스템 API
  version: 1.0.0
  contact:
    name: API Support
    email: support@golfadmin.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.golfadmin.com
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /tee-times:
    get:
      tags:
        - TeeTimes
      summary: 티타임 목록 조회
      operationId: getTeeTimesList
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: golfCourseId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TeeTimeStatus'
        - name: teeTimeType
          in: query
          schema:
            $ref: '#/components/schemas/TeeTimeType'
        - name: bookingType
          in: query
          schema:
            $ref: '#/components/schemas/BookingType'
      responses:
        '200':
          description: 티타임 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeeTimeListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - TeeTimes
      summary: 티타임 등록
      operationId: createTeeTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeeTimeRequest'
      responses:
        '201':
          description: 티타임 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeeTime'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /tee-times/matrix:
    get:
      tags:
        - TeeTimes
      summary: 매트릭스 뷰 데이터 조회
      operationId: getTeeTimesMatrix
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TeeTimeType'
          required: false
        - name: booking
          in: query
          schema:
            $ref: '#/components/schemas/BookingType'
          required: false
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          required: false
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 90
      responses:
        '200':
          description: 매트릭스 데이터
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatrixViewResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /tee-times/reserve:
    post:
      tags:
        - TeeTimes
      summary: 티타임 예약 (10분 타이머 시작)
      operationId: reserveTeeTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveTeeTimeRequest'
      responses:
        '200':
          description: 예약 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReserveTeeTimeResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
  /tee-times/timer-status/{id}:
    get:
      tags:
        - TeeTimes
      summary: 예약 타이머 상태 조회
      operationId: getTimerStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 타이머 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 페이지 번호 (1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: 정렬 기준 (field:asc 또는 field:desc)
      schema:
        type: string
        pattern: '^[a-zA-Z]+:(asc|desc)$'
  schemas:
...
    AccountType:
      type: string
      enum:
        - SUPER_ADMIN
        - ADMIN
        - TEAM_LEADER
        - INTERNAL_MANAGER
        - EXTERNAL_MANAGER
        - PARTNER
        - GOLF_COURSE
        - MEMBER
    Region:
      type: string
      enum:
        - 제주
        - 경기
        - 강원
        - 충청
        - 호남
        - 영남
        - 경상
        - 기타
    TeeTimeStatus:
      type: string
      enum:
        - AVAILABLE
        - RESERVED
        - CONFIRMED
        - COMPLETED
        - CANCELLED
    TeeTimeType:
      type: string
      enum:
        - DAILY
        - PACKAGE
    BookingType:
      type: string
      enum:
        - BOOKING
        - JOIN
    TimeSlot:
      type: string
      enum:
        - 1부
        - 2부
        - 3부
    GolfCourseStatus:
      type: string
      enum:
        - ACTIVE
        - API_INTEGRATION
        - MANUAL
        - INACTIVE

    # Error Response
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
        details:
          type: object
          description: 추가 에러 상세 정보
    ValidationError:
      type: object
      required:
        - code
        - message
        - errors
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        name:
          type: string
        accountType:
          $ref: '#/components/schemas/AccountType'
        teamId:
          type: string
          nullable: true
        team:
          $ref: '#/components/schemas/Team'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        leaderId:
          type: string
        leader:
          $ref: '#/components/schemas/User'
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GolfCourse:
      type: object
      properties:
        id:
          type: string
        order:
          type: integer
        region:
          $ref: '#/components/schemas/Region'
        name:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/GolfCourseStatus'
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TeeTime:
      type: object
      properties:
        id:
          type: string
        golfCourseId:
          type: string
        golfCourse:
          $ref: '#/components/schemas/GolfCourse'
        date:
          type: string
          format: date
        time:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
        timeSlot:
          $ref: '#/components/schemas/TimeSlot'
        greenFee:
          type: number
          format: decimal
          description: 그린피 (만원 단위, 소수점 1자리)
        playerCount:
          type: integer
          minimum: 1
          maximum: 4
        requirements:
          type: string
        holeType:
          type: string
          enum: [18홀, 9홀]
        caddie:
          type: boolean
        advance:
          type: number
          format: decimal
          description: 선입금
        mealIncluded:
          type: boolean
        cartIncluded:
          type: boolean
        teeTimeType:
          $ref: '#/components/schemas/TeeTimeType'
        bookingType:
          $ref: '#/components/schemas/BookingType'
        status:
          $ref: '#/components/schemas/TeeTimeStatus'
        reservedAt:
          type: string
          format: date-time
          nullable: true
        managerId:
          type: string
          nullable: true
        manager:
          $ref: '#/components/schemas/User'
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        confirmedBy:
          type: string
          nullable: true
        connectedTeeTimeId:
          type: string
          nullable: true
          description: 연결된 티타임 ID (패키지용)
        accommodationInfo:
          type: string
          nullable: true
          description: 숙박 정보 (패키지용)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTeeTimeRequest:
      type: object
      required:
        - golfCourseId
        - date
        - time
        - greenFee
        - playerCount
        - requirements
        - holeType
        - caddie
        - advance
        - mealIncluded
        - cartIncluded
      properties:
        golfCourseId:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
        greenFee:
          type: number
          format: decimal
        playerCount:
          type: integer
          minimum: 1
          maximum: 4
        requirements:
          type: string
        holeType:
          type: string
          enum: [18홀, 9홀]
        caddie:
          type: boolean
        advance:
          type: number
          format: decimal
        mealIncluded:
          type: boolean
        cartIncluded:
          type: boolean
        teeTimeType:
          $ref: '#/components/schemas/TeeTimeType'
          default: DAILY
        connectedTeeTimeId:
          type: string
        accommodationInfo:
          type: string
...
    TeeTimeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TeeTime'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ReserveTeeTimeRequest:
      type: object
      required:
        - teeTimeId
      properties:
        teeTimeId:
          type: string
    ReserveTeeTimeResponse:
      type: object
      properties:
        teeTime:
          $ref: '#/components/schemas/TeeTime'
        expiresAt:
          type: string
          format: date-time
          description: 예약 만료 시간 (10분 후)
...
    TimerStatusResponse:
      type: object
      properties:
        teeTimeId:
          type: string
        status:
          $ref: '#/components/schemas/TeeTimeStatus'
        reservedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        remainingSeconds:
          type: integer
          description: 남은 시간 (초)
        isExpired:
          type: boolean

    # Matrix View Schemas
    MatrixViewResponse:
      type: object
      properties:
        matrixData:
          type: array
          items:
            $ref: '#/components/schemas/RegionMatrix'
        dateColumns:
          type: array
          items:
            $ref: '#/components/schemas/DateColumn'
        summary:
          $ref: '#/components/schemas/MatrixSummary'
    RegionMatrix:
      type: object
      properties:
        region:
          $ref: '#/components/schemas/Region'
        golfCourses:
          type: array
          items:
            $ref: '#/components/schemas/GolfCourseMatrix'
    GolfCourseMatrix:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        dates:
          type: array
          items:
            $ref: '#/components/schemas/DateMatrix'
    DateMatrix:
      type: object
      properties:
        date:
          type: string
          format: date
        timeSlot1:
          type: integer
          description: 1부 티타임 수
        timeSlot2:
          type: integer
          description: 2부 티타임 수
        timeSlot3:
          type: integer
          description: 3부 티타임 수
        total:
          type: integer
          description: 전체 티타임 수
    DateColumn:
      type: object
      properties:
        date:
          type: string
          format: date
        displayDate:
          type: string
          description: 표시용 날짜 (MM/DD)
        dayOfWeek:
          type: string
          description: 요일
        isToday:
          type: boolean
        isWeekend:
          type: boolean
    MatrixSummary:
      type: object
      properties:
        totalGolfCourses:
          type: integer
        totalTeeTimes:
          type: integer
        teeTimeType:
          $ref: '#/components/schemas/TeeTimeType'
        bookingType:
          $ref: '#/components/schemas/BookingType'
        dateRange:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date

    # Performance Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
  responses:
    BadRequestError:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    UnauthorizedError:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: 인증이 필요합니다
    ForbiddenError:
      description: 권한 부족
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: 해당 작업을 수행할 권한이 없습니다
    NotFoundError:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: 요청한 리소스를 찾을 수 없습니다
    ConflictError:
      description: 충돌 발생
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: 리소스 충돌이 발생했습니다
...